refactor: move Logger types to types/internal and reexport from logger pkg
