+++ packages/gateway/src/interfaces/Config.ts
@@ -1,4 +1,3 @@
-import type { LoggerConfig } from '@seedcord/logger';
 import type { BotConfig, Config } from '@seedcord/types';
 import type { ClientOptions } from 'discord.js';
 
@@ -39,10 +38,4 @@ export interface GatewayBotConfig extends BotConfig {
  */
 export interface GatewayConfig extends Config {
     bot: GatewayBotConfig;
-
-    /**
-     * Logging level, sinks, and per-channel overrides. Omitted fields keep the node defaults (a pretty
-     * console plus a combined file in dev, JSON to the console in prod).
-     */
-    logger?: LoggerConfig;
 }
