+++ packages/types/src/internal.index.ts
@@ -1,5 +1,6 @@
 export * from './brand';
 export type * from './Hmr';
 export { wrapHot } from './Hmr';
+export type * from './Interfaces/LogSink';
 export type * from './Interfaces/SeedcordInstance';
 export type * from './Interfaces/Store';
